chore(pkg): Refactor Pkg Skeleton
