+++ resources/out_key.txt
@@ -1,0 +1,4 @@
+a
+b
+c
+d
