+++ resources/out_value.txt
@@ -1,0 +1,4 @@
+1
+2
+3
+4
