+++ src/network/server.cpp
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <unordered_map>
-#include "engine/db.h" 
+#include "storage/old.h" 
 #include "grpcpp/grpcpp.h"
-#include "service.grpc.pb.h"
+#include "kvservice.grpc.pb.h"
 #include <glog/logging.h>
-#include "engine/types/leveldb_engine.h"
-#include "engine/types/plainkv_engine.h"
+#include "storage/engine/leveldb_engine.h"
+#include "storage/engine/plainkv_engine.h"
 using grpc::Server;
 using grpc::ServerBuilder;
 using grpc::ServerContext;
